# Return bools from archive/delete_batch