--- a/sql/pgmq--0.21.0--0.22.0.sql
+++ b/sql/pgmq--0.21.0--0.22.0.sql
@@ -1,0 +1,28 @@
+ALTER EXTENSION pgmq DROP FUNCTION pgmq_delete(TEXT, bigint[]);
+DROP FUNCTION pgmq_delete(TEXT, bigint[]);
+
+-- src/lib.rs:215
+-- pgmq::pgmq_delete
+CREATE FUNCTION "pgmq_delete"(
+	"queue_name" TEXT, /* &str */
+	"msg_ids" bigint[] /* alloc::vec::Vec<i64> */
+) RETURNS TABLE (
+	"pgmq_delete" bool  /* bool */
+)
+STRICT 
+LANGUAGE c /* Rust */
+AS 'MODULE_PATHNAME', 'pgmq_delete_batch_wrapper';
+
+ALTER EXTENSION pgmq ADD FUNCTION pgmq_delete(TEXT, bigint[]);
+
+-- -- src/lib.rs:260
+-- -- pgmq::pgmq_archive
+-- CREATE  FUNCTION "pgmq_archive"(
+-- 	"queue_name" TEXT, /* &str */
+-- 	"msg_ids" bigint[] /* alloc::vec::Vec<i64> */
+-- ) RETURNS TABLE (
+-- 	"pgmq_archive" bool  /* bool */
+-- )
+-- STRICT 
+-- LANGUAGE c /* Rust */
+-- AS 'MODULE_PATHNAME', 'pgmq_archive_batch_wrapper';