--- a/sql/pgmq--0.21.0--0.22.0.sql
+++ b/sql/pgmq--0.21.0--0.22.0.sql
@@ -13,16 +13,18 @@ STRICT
 LANGUAGE c /* Rust */
 AS 'MODULE_PATHNAME', 'pgmq_delete_batch_wrapper';
 
-ALTER EXTENSION pgmq ADD FUNCTION pgmq_delete(TEXT, bigint[]);
 
--- -- src/lib.rs:260
--- -- pgmq::pgmq_archive
--- CREATE  FUNCTION "pgmq_archive"(
--- 	"queue_name" TEXT, /* &str */
--- 	"msg_ids" bigint[] /* alloc::vec::Vec<i64> */
--- ) RETURNS TABLE (
--- 	"pgmq_archive" bool  /* bool */
--- )
--- STRICT 
--- LANGUAGE c /* Rust */
--- AS 'MODULE_PATHNAME', 'pgmq_archive_batch_wrapper';
+ALTER EXTENSION pgmq DROP FUNCTION pgmq_archive(TEXT, bigint[]);
+DROP FUNCTION pgmq_archive(TEXT, bigint[]);
+
+-- src/lib.rs:260
+-- pgmq::pgmq_archive
+CREATE  FUNCTION "pgmq_archive"(
+	"queue_name" TEXT, /* &str */
+	"msg_ids" bigint[] /* alloc::vec::Vec<i64> */
+) RETURNS TABLE (
+	"pgmq_archive" bool  /* bool */
+)
+STRICT 
+LANGUAGE c /* Rust */
+AS 'MODULE_PATHNAME', 'pgmq_archive_batch_wrapper';