Closes #89.

This is an attempt at resolving #89. I also refactored archive in terms of archive_batch. Please let me know if you want me to revert that part.

It has been quite a while since I have written any Rust. I get the feeling I should be doing more `map`ing and `and_then`ing and using less `mut` 🙂 

Before
```
pgmq=# select * from pgmq_archive('poop', array[1, 3, 5]);
 pgmq_archive
--------------
 t
(1 row)

pgmq=# select * from pgmq_archive('poop', array[1, 3, 5]);
 pgmq_archive
--------------
 t
(1 row)
```

After
```
pgmq=# select * from pgmq_archive('poop', array[1, 3, 5]);
 pgmq_archive 
----------
 t
 t
 f
(3 rows)

pgmq=# select * from pgmq_archive('poop', array[1, 3, 5]);
 pgmq_archive 
----------
 f
 f
 f
(3 rows)
```