@craigpastro 2023-08-30
So I wrote a small script that looks like
```
select pgmq_create('q');
select * from pgmq_send('q', '{"foo":"bar"}');
select * from pgmq_archive('q', 1);
... (a bunch more)
select pgmq_drop_queue('q');
```
and ran pgbench with that. The results are below. There doesn't seem to be much of a difference, but I am not an expert of pgbench or Postgres benchmarks. Any suggestions? Thanks!


Current
```
$ pgbench -h localhost -p 28815 -U craig --client=1 --jobs=1 -t 500 -n --file=pgmq_bench.sql pgmq
pgbench (15.3 (Homebrew))
transaction type: pgmq_bench.sql
scaling factor: 1
query mode: simple
number of clients: 1
number of threads: 1
maximum number of tries: 1
number of transactions per client: 500
number of transactions actually processed: 500/500
number of failed transactions: 0 (0.000%)
latency average = 8.198 ms
initial connection time = 5.505 ms
tps = 121.983560 (without initial connection time)
```

This PR
```
$ pgbench -h localhost -p 28815 -U craig --client=1 --jobs=1 -t 500 -n --file=pgmq_bench.sql pgmq
pgbench (15.3 (Homebrew))
transaction type: pgmq_bench.sql
scaling factor: 1
query mode: simple
number of clients: 1
number of threads: 1
maximum number of tries: 1
number of transactions per client: 500
number of transactions actually processed: 500/500
number of failed transactions: 0 (0.000%)
latency average = 8.246 ms
initial connection time = 4.762 ms
tps = 121.263713 (without initial connection time)
```