@ChuckHend 2023-09-01
The `in` is about 0.01 ms slower than the `=` according this bench, just comparing the deletes. I'm curious if there are situations that could make that better/worse.


```text
pgbench postgres://user:abc@localhost:28815/pgmq -P 1 -r -T 60 -f in.sql -f eq.sql
transaction type: multiple scripts
scaling factor: 1
query mode: simple
number of clients: 1
number of threads: 1
maximum number of tries: 1
duration: 60 s
number of transactions actually processed: 102134
number of failed transactions: 0 (0.000%)
latency average = 0.587 ms
latency stddev = 0.120 ms
initial connection time = 3.283 ms
tps = 1702.194694 (without initial connection time)
SQL script 1: in.sql
 - weight: 1 (targets 50.0% of total)
 - 51290 transactions (50.2% of total, tps = 854.813929)
 - number of failed transactions: 0 (0.000%)
 - latency average = 0.593 ms
 - latency stddev = 0.146 ms
 - statement latencies in milliseconds and failures:
         0.593           0  delete from bencher where record_id in (select max(record_id) from bencher)
SQL script 2: eq.sql
 - weight: 1 (targets 50.0% of total)
 - 50844 transactions (49.8% of total, tps = 847.380764)
 - number of failed transactions: 0 (0.000%)
 - latency average = 0.582 ms
 - latency stddev = 0.086 ms
 - statement latencies in milliseconds and failures:
         0.582           0  delete from bencher where record_id = (select max(record_id) from bencher)
```