@craigpastro 2023-09-01
🤔 it's unlikely, but archive_batch also has to go through some string construction and manipulation which could add a slight latency. Let me try to replace `IN` with `= ANY` to avoid the string manipulation (and it is also nicer I think) and can you try to run the benchmark again?