@craigpastro 2023-09-01
I tried explain analyze just to see what happens when using `= 1` and `= ANY(array[1])` and the plan is the same except for the index condition, but I can't see that having much of an impact. It just so happens here that the former was a bit slower, but those times fluctuate quite a bit.

```
pgmq=# EXPLAIN ANALYZE SELECT * FROM pgmq_q WHERE msg_id = 1;
                                                     QUERY PLAN                                                      
---------------------------------------------------------------------------------------------------------------------
 Index Scan using pgmq_q_pkey on pgmq_q  (cost=0.15..8.17 rows=1 width=60) (actual time=0.019..0.020 rows=0 loops=1)
   Index Cond: (msg_id = 1)
 Planning Time: 0.236 ms
 Execution Time: 0.130 ms
(4 rows)

pgmq=# EXPLAIN ANALYZE SELECT * FROM pgmq_q WHERE msg_id = ANY(array[1]);
                                                     QUERY PLAN                                                      
---------------------------------------------------------------------------------------------------------------------
 Index Scan using pgmq_q_pkey on pgmq_q  (cost=0.15..8.17 rows=1 width=60) (actual time=0.023..0.024 rows=0 loops=1)
   Index Cond: (msg_id = ANY ('{1}'::integer[]))
 Planning Time: 0.224 ms
 Execution Time: 0.058 ms
(4 rows)
```


What was kind of interesting is that `in (...)` gets transformed into `= ANY(...)`, or at least this is how I interpret below:
```
pgmq=# EXPLAIN ANALYZE SELECT * FROM pgmq_q WHERE msg_id in (1,2,3);
                                                     QUERY PLAN                                                     
--------------------------------------------------------------------------------------------------------------------
 Bitmap Heap Scan on pgmq_q  (cost=4.47..11.58 rows=3 width=60) (actual time=0.009..0.010 rows=0 loops=1)
   Recheck Cond: (msg_id = ANY ('{1,2,3}'::bigint[]))
   ->  Bitmap Index Scan on pgmq_q_pkey  (cost=0.00..4.47 rows=3 width=0) (actual time=0.005..0.006 rows=0 loops=1)
         Index Cond: (msg_id = ANY ('{1,2,3}'::bigint[]))
 Planning Time: 0.459 ms
 Execution Time: 0.112 ms
(6 rows)
```