@ChuckHend 2023-09-02
Will need to bump core/Cargo.toml and ./Cargo.toml versions.

Also add an empty migration script for this version.