@ChuckHend 2023-09-02
> 👋 @ChuckHend. Bumped minor versions and added empty migration scripts. I added 0.21 to 0.22 too since we are now at 0.22. Is that correct? Thanks!

Yes! Good call. It's easy to forget those. Wish we had CI checks for them.